# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Runic API
  description: |
    Firewall rule management system API.
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Error Responses
    All endpoints may return error responses in the following format:
    ```json
    {
      "error": "error message"
    }
    ```
    
    Common HTTP status codes:
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Missing or invalid token
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 409: Conflict - Resource in use
    - 500: Internal Server Error
  version: 1.0.0
  contact:
    name: Runic Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: /api/v1
    description: Production server (relative path)
...
paths:
...
  /groups:
    get:
      tags:
        - Groups
      summary: List all groups
      description: Returns a list of all groups with peer and policy counts.
      operationId: listGroups
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Groups
      summary: Create a new group
      description: Creates a new group.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
  /groups/{id}:
    get:
      tags:
        - Groups
      summary: Get group by ID
      description: Returns detailed information about a specific group.
      operationId: getGroup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    put:
      tags:
        - Groups
      summary: Update group
      description: Updates an existing group.
      operationId: updateGroup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "group updated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Groups
      summary: Delete group
      description: Deletes a group.
      operationId: deleteGroup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Group deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Group deleted"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
  /groups/{id}/members:
    get:
      tags:
        - Groups
      summary: List group members
      description: Returns a list of peers that are members of the group.
      operationId: listGroupMembers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Peer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Groups
      summary: Add member to group
      description: Adds a peer to a group.
      operationId: addGroupMember
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - peer_id
              properties:
                peer_id:
                  type: integer
                  format: int64
                  description: The ID of the peer to add to the group
      responses:
        '200':
          description: Member added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "member added"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    Peer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hostname:
          type: string
        ip_address:
          type: string
          format: ipv4
        os_type:
          type: string
          enum: [debian, ubuntu, rhel, arch, opensuse, raspbian, linux]
        arch:
          type: string
          enum: [amd64, arm64, arm, armv6]
        has_docker:
          type: boolean
        is_manual:
          type: boolean
        agent_version:
          type: string
        last_heartbeat:
          type: string
          format: date-time
        groups:
          type: string
          description: Comma-separated group names
        status:
          type: string
          enum: [online, offline, pending]
        bundle_version:
          type: string
        bundle_version_number:
          type: integer
        description:
          type: string
        hmac_key_last_rotated_at:
          type: string
          format: date-time
        pending_changes_count:
          type: integer
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
        peer_count:
          type: integer
        policy_count:
          type: integer
    CreateGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Group name (1-100 chars, alphanumeric with underscore/hyphen/space)
        description:
          type: string
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string

    # Services
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'